+++ .changeset/sixty-eels-camp.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/db": patch
+---
+
+Prevent runtime from importing core code
