+++ packages/db/src/core/errors.ts
@@ -37,44 +37,12 @@ export const SHELL_QUERY_MISSING_ERROR = `${red(
 	'▶ Please provide a query to execute using the --query flag.'
 )}\n`;
 
-export const SEED_ERROR = (error: string) => {
-	return `${red(`Error while seeding database:`)}\n\n${error}`;
-};
-
 export const EXEC_ERROR = (error: string) => {
 	return `${red(`Error while executing file:`)}\n\n${error}`;
 };
 
-export const SEED_DEFAULT_EXPORT_ERROR = (fileName: string) => {
-	return SEED_ERROR(`Missing default function export in ${bold(fileName)}`);
-};
-
 export const EXEC_DEFAULT_EXPORT_ERROR = (fileName: string) => {
 	return EXEC_ERROR(`Missing default function export in ${bold(fileName)}`);
-};
-
-export const REFERENCE_DNE_ERROR = (columnName: string) => {
-	return `Column ${bold(
-		columnName
-	)} references a table that does not exist. Did you apply the referenced table to the \`tables\` object in your db config?`;
-};
-
-export const FOREIGN_KEY_DNE_ERROR = (tableName: string) => {
-	return `Table ${bold(
-		tableName
-	)} references a table that does not exist. Did you apply the referenced table to the \`tables\` object in your db config?`;
-};
-
-export const FOREIGN_KEY_REFERENCES_LENGTH_ERROR = (tableName: string) => {
-	return `Foreign key on ${bold(
-		tableName
-	)} is misconfigured. \`columns\` and \`references\` must be the same length.`;
-};
-
-export const FOREIGN_KEY_REFERENCES_EMPTY_ERROR = (tableName: string) => {
-	return `Foreign key on ${bold(
-		tableName
-	)} is misconfigured. \`references\` array cannot be empty.`;
 };
 
 export const INTEGRATION_TABLE_CONFLICT_ERROR = (
