+++ packages/db/src/runtime/seed-local.ts
@@ -2,7 +2,7 @@ import { LibsqlError } from '@libsql/client';
 import { type SQL, sql } from 'drizzle-orm';
 import type { LibSQLDatabase } from 'drizzle-orm/libsql';
 import { SQLiteAsyncDialect } from 'drizzle-orm/sqlite-core';
-import { SEED_DEFAULT_EXPORT_ERROR, SEED_ERROR } from '../core/errors.js';
+import { SEED_DEFAULT_EXPORT_ERROR, SEED_ERROR } from './errors.js';
 import { type DBTables } from '../core/types.js';
 import { getCreateIndexQueries, getCreateTableQuery } from './queries.js';
